\name{ccr81}
\docType{data}
\alias{ccr81}

\title{Program Follow Through at Primary Schools}

\description{
 The data set is from an US federally sponsored program for providing remedial assistance to disadvantaged primary school students. The data comprises 70 school sites.
}

\usage{data( ccr81 )}

\format{
	This data frame contains the following variables (columns):
  \describe{
    \item{\code{nu}}{School Site Number}    		\item{\code{y1}}{Total Reading Score}
    \item{\code{y2}}{Total Math Score}
    \item{\code{y3}}{Total Coopersmith Score}

    \item{\code{x1}}{Education Level of Mother}
    \item{\code{x2}}{Occupation Index}
    \item{\code{x3}}{Parental Visit Index}
    \item{\code{x4}}{Counseling Index}
    \item{\code{x5}}{Number of Teachers}
  }
}


\details{
	The data were originally used to evaluate the efficiency of public programs and their management.
}

\source{
A. Charnes, W. W. Cooper and E. Rhodes (1981), Evaluating Program and Managerial Efficiency: An Application of Data Envelopment Analysis to Program Follow Through, \emph{Management Science}, \bold{27}, 668--697.
}

\references{
Charnes, A., W. W. Cooper, and E. Rhodes. 1981. Evaluating Program and Managerial Efficiency: An Application of Data Envelopment Analysis to Program Follow Through. \emph{Management Science} \bold{27}: 668--697
}

\keyword{datasets}
