\name{mroz}
\docType{data}
\alias{mroz}

\title{Female labor force participation}

\description{
 Instructional dataset, N=753, cross-sectional labor force participation data Accompanying Introductory Econometrics: A Modern Approach, Jeffrey M. Wooldridge, South-Western College Publishing, (c) 2000 and Jeffrey M. Wooldridge, Econometric Analysis of Cross Section and Panel Data, MIT Press,(c) 2001.
}

\usage{data( mroz )}

\format{
	This data frame contains the following variables (columns):
  \describe{
    \item{\code{inlf}}{=1 if in labor force, 1975}
    \item{\code{hours}}{hours worked, 1975}
    \item{\code{kidslt6}}{# kids < 6 years}
    \item{\code{kidsge6}}{# kids 6-18}
    \item{\code{age}}{woman's age in yrs}
    \item{\code{educ}}{years of schooling}
    \item{\code{wage}}{estimated wage from earns., hours}
    \item{\code{repwage}}{reported wage at interview in 1976}
    \item{\code{hushrs}}{hours worked by husband, 1975}
    \item{\code{husage}}{husband's age}
    \item{\code{huseduc}}{husband's years of schooling}
    \item{\code{huswage}}{husband's hourly wage, 1975}
    \item{\code{faminc}}{family income, 1975}
    \item{\code{mtr}}{fed. marginal tax rate facing woman}
    \item{\code{motheduc}}{mother's years of schooling}
    \item{\code{fatheduc}}{father's years of schooling}
    \item{\code{unem}}{unem. rate in county of resid.}
    \item{\code{city}}{=1 if live in SMSA}
    \item{\code{exper}}{actual labor mkt exper}
    \item{\code{nwifeinc}}{(faminc - wage*hours)/1000}
  }
}


\details{
	Instructional dataset, N=753, cross-sectional labor force participation data Accompanying Introductory Econometrics: A Modern Approach, Jeffrey M. Wooldridge, South-Western College Publishing, (c) 2000 and Jeffrey M. Wooldridge, Econometric Analysis of Cross Section and Panel Data, MIT Press,(c) 2001.
}

\source{
Datasets accessible from http://wooldridge.swcollege.com
}

\references{
Mroz, T.A. 1987. The Sensitiviy of an Empirical Model of Married Women's Hours of Work to Economic and Statistical Assumptions. \emph{Econometrica} \bold{55}: 765-799.
}

\keyword{datasets}
